requirements.txt
langchain
openai
faiss-cpu
chainlit
requests
python-dotenv
